Switch from Eleventy to Parcel (#3)
I realized it was premature to use Eleventy, where the motivation is
separating buildtime from runtime. A lot of the complexity will actually
be at runtime to dynamically set up the form. So, I'm going to focus on
building the app as simple/maintainable as possible using hardcoded
config, and only later generalize for the config to be less hardcoded.

I also realized that there will be other ways in the future to update
the config than needing to use Eleventy. For example, we can use a JSON
file and a script that updates that JSON file from Google Tables. Then
we distribute the JSON file with the app. This is what
https://github.com/ParkingReformNetwork/reform-map does.

This uses https://parceljs.org, which works great to generate the site
and automatically handle TypeScript and Sass. The PR also sets up
Playwright tests.