Determine specific errors in link generation (#14)
This new paradigm is much better. We now result something like Rust's
`Result`, whereas before we turned the equivalent of `Option`.

This allows us to:

* Give more useful error messages
* Have >1 error
* Have errors specific to each communication type, e.g. a question may
be required in one but not the other

We don't yet surface these errors to the end user, but that will be a
follow up.