Add support for optional fields (#21)
You must actively choose the option "Do not set this value". This solves
two goals:

1. Reduce the risk of not setting the value when you should. Force the
user to make a choice
2. Provide a way to reset the radio group. If you accidentally chose an
actual value but wanted to leave it blank, we need some way to reset to
"no selection". I could have added a reset button, but it's easier to
have an "n/a" radio button.

Note that we don't have to do anything special in `initForm.ts`, like
converting the None option to `undefined`. We want to preserve the raw
value of the `None` option and pass it to `linkGenerator.ts` because it
needs a way to distinguish between a fieldset being completely
unanswered versus being marked as none.