Refactor result container creation (#36)
We now set up the HTML for the containers and overall structure eagerly
in `index.html` and only update individual nodes. This will make it
easier to set up the copy link button and its event listener so that we
only create the node once and don't have to re-create it every time the
URL changes.